Creating custom role to access storage bucket and configuring auth micro-service to use storage module

+++ libs/common/src/storage/storage.service.ts
@@ -9,7 +9,9 @@ import { StreamOutput } from './interfaces/stream-output.interface';
 
 @Injectable()
 export class StorageService {
-  constructor(@Inject('BUCKET') private readonly storageBucket: Bucket) {}
+  constructor(
+    @Inject('BUCKET') private readonly storageBucket: Bucket | undefined,
+  ) {}
 
   async save(
     path: string,
@@ -17,48 +19,60 @@ export class StorageService {
     media: Buffer,
     metadata: { [key: string]: string }[],
   ) {
-    return new Promise<StreamOutput>((resolve, reject) => {
-      const object = metadata.reduce(
-        (obj, item) => Object.assign(obj, item),
-        {},
-      );
+    if (this.storageBucket) {
+      return new Promise<StreamOutput>((resolve, reject) => {
+        const object = metadata.reduce(
+          (obj, item) => Object.assign(obj, item),
+          {},
+        );
 
-      const file = this.storageBucket.file(path);
-      const stream = file.createWriteStream({
-        metadata: {
-          contentType,
-        },
-      });
-
-      stream
-        .on('error', (error) => {
-          reject(new InternalServerErrorException(error.message));
-        })
-        .on('finish', async () => {
-          try {
-            await file.setMetadata({
-              metadata: object,
-            });
-            await file.makePublic();
-
-            const streamOutput: StreamOutput = {
-              publicUrl: `https://storage.googleapis.com/${this.storageBucket.name}/${path}`,
-            };
-            resolve(streamOutput);
-          } catch (error) {
-            reject(new InternalServerErrorException(error.message));
-          }
+        const file = this.storageBucket.file(path);
+        const stream = file.createWriteStream({
+          metadata: {
+            contentType,
+          },
         });
 
-      stream.end(media);
-    });
+        stream
+          .on('error', (error) => {
+            reject(new InternalServerErrorException(error.message));
+          })
+          .on('finish', async () => {
+            try {
+              await file.setMetadata({
+                metadata: object,
+              });
+              await file.makePublic();
+
+              const streamOutput: StreamOutput = {
+                publicUrl: `https://storage.googleapis.com/${this.storageBucket.name}/${path}`,
+              };
+              resolve(streamOutput);
+            } catch (error) {
+              reject(new InternalServerErrorException(error.message));
+            }
+          });
+
+        stream.end(media);
+      });
+    } else {
+      throw InternalServerErrorException;
+    }
   }
 
   async delete(path: string) {
-    return this.storageBucket.file(path).delete({ ignoreNotFound: true });
+    if (this.storageBucket) {
+      return this.storageBucket.file(path).delete({ ignoreNotFound: true });
+    } else {
+      throw InternalServerErrorException;
+    }
   }
 
   async deleteFilesByDirName(path: string) {
-    return this.storageBucket.deleteFiles({ prefix: path });
+    if (this.storageBucket) {
+      return this.storageBucket.deleteFiles({ prefix: path });
+    } else {
+      throw InternalServerErrorException;
+    }
   }
 }
